requirements.txt
google-adk[extensions]>=1.0
python-dotenv>=1.0
